# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Blink HTTP API
  description: |
    本仓库 HTTP 接口的 OpenAPI 描述；与 `infrastructure/interface/http/` 及 `cmd/main.go` 中的路由一致。

    **`POST /auth/register`**：始终挂载（见 `cmd/main.go`），用于邮箱密码注册并建立 `builtin` 身份绑定。

    **内置 IdP**：`/auth/idp/*` 以及 OAuth 提供方 **`builtin`** 仅在配置了 `BLINK_PUBLIC_BASE_URL`、`BLINK_OAUTH_CLIENT_SECRET` 等环境变量时启用；未配置时上述路径返回 **404**，但不影响 `/auth/register`。

    **代码生成**：仅带标签 `servergen` 的操作会由 oapi-codegen 生成 Gin 注册代码与 `ServerInterface` 方法；其余路径为手写路由，此处仅作文档。

    **社交与管理**：`/api/*`、`/admin/api/*` 为手写路由（见 `cmd/main.go`），OpenAPI 中 `social` / `admin` / `feedback` / `notifications` 标签描述之。

    **管理权限**：`users.role` 为 `admin` 或 `super_admin` 时可访问 `/admin/api/*`（需有效会话）。部分能力（如 SMTP 配置）在实现上仍建议仅授予 `super_admin`。
  version: 0.1.0
servers:
  - url: http://127.0.0.1:11110
    description: 本地开发（默认 BLINK_HTTP_ADDR=:11110）
  - url: http://localhost:11110
    description: 本地开发（localhost）
  - url: "http://127.0.0.1:{port}"
    description: 本地其它端口（与当前 BLINK_HTTP_ADDR 一致，在 Swagger 里改 port）
    variables:
      port:
        default: "11110"
  - url: /
    description: 与 Swagger UI 同源（把文档托管在本 API 上时用此项）
...
paths:
...
  /api/me/posts/following:
    get:
      operationId: listFollowingPosts
      tags: [social]
      summary: 关注流（需登录）
      security:
        - sessionCookie: []
        - bearerAuth: []
      description: |
        返回当前用户所关注用户发布的公开已发布原创帖，按 id 降序分页。
        Query：`cursor`（上一页最后一条帖子 id）、`limit`（默认 20，最大 50）。
      responses:
        "200":
          description: 同 listPosts
        "401":
          description: 未登录
  /api/search/posts:
    get:
      operationId: searchPosts
      tags: [social]
      summary: 搜索帖子（可选登录）
      description: |
        Query：`q`（必填，1–100 字符）、`cursor`、`limit`。
        仅匹配已发布、公开、审核正常的原创帖正文；登录时含 `liked`。
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: 同 listPosts
        "400":
          description: 缺少或无效 q
  /api/search/users:
    get:
      operationId: searchUsers
      tags: [social]
      summary: 搜索用户（按展示名）
      description: |
        Query：`q`（必填）、`offset`（默认 0）、`limit`（默认 20，最大 50）。
        仅返回 status=正常 的用户；响应不含邮箱。
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: offset
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: '`{ "users": [...], "next_offset": int|null }`'
        "400":
          description: 缺少或无效 q
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: access_token
    sessionCookie:
      type: apiKey
      in: cookie
      name: blink_session
...